Fixed Version
I had accidentally made the End of Box Minigame not working by dropping a gameObject onto Level1_Finalized>Minigames>BoxMinigames>Door Blockables.
This caused the level to not complete, and to isolate the issue, I had to revert to a working version of the commits history.
EVERYTHING IS FINE AND WORK AS BEFORE NOW!